Show sample output from program
The example code here has an elaborate set up. If you're looking to learn how Semaphore works, you're unlikely to know what this code is supposed to do.

Including a sample of the program's output immediately after the code allows the reader to make deeper intuitions about the behaviour of the Semaphore class.